% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.reditR_validation}
\alias{print.reditR_validation}
\title{Print a reditR validation summary}
\usage{
\method{print}{reditR_validation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"reditR_validation"}.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a reditR validation summary
}
